Completed Exercises 4.18-4.23

+++ part4/blog/controllers/blogsRouter.js
@@ -3,6 +3,10 @@ const Blog = require('../models/blog')
 const logger = require('../utils/logger')
 require('express-async-errors')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -11,15 +15,13 @@ blogsRouter.get('/', async (request, response) => {
 
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   const body = request.body
+  const user = request.user
 
   if (!body.title || !body.url) {
     return response.status(400).json({ error: "title or url is missing"})
   }
-
-  // const user = await User.findById(body.userId)
-  const user = await User.findOne({})
 
   const blog = new Blog({
     title: body.title,
@@ -37,7 +39,14 @@ blogsRouter.post('/', async (request, response, next) => {
 
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(400).json({ error: 'user is not blog creator'})
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
